DigitSum and Palindrome Questions
I completed

DigitSum Question:
Create a program that adds up the individual digits that make up a number greater then 10

Palindrome:
Create a boolean method to determine if a number is a palindrome (1221: true, 1222: false)

I used modulo 10 to deduct the end of the input, then ADDED that to a 'reverse' number to build it backwards. Then compared it to the original number to see if they are equal.